Bike sharing patterns
This project predicts bike sharing patterns in a city. Given features like day of the week, weather, temperature, holidays and others, the neural network model predicts how many bikes should a bike store keep in a given day such that they are not siting idle in the store, and also are enough to meet the demand.